feat(fields): do not pass NaN in FinF [CLK-656138]
If a formula is a variable to another formula, we can't use NaN to indicate the impossibility to calculate it (for example because of a missing variable value). In such cases, we must provide null to indicate missing value as with normal variables.

Fix added and a test case for it.